bloglist expansion, step11

+++ part4/blog-list/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/', async (request, response) => {
   blogs = await Blog
@@ -10,14 +11,19 @@ blogRouter.get('/', async (request, response) => {
 
 blogRouter.post('/', async (request, response) => {
   const body = request.body
-  const user = await User.findById(body.userId)
+  
+  if (!request.token)
+    return response.status(401).json({ error: 'token missing' })
+  jwt.verify(request.token, process.env.SECRET)
+
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
-    author: body.author,
+    author: user.name || body.author,
     url: body.url,
     likes: body.likes,
-    user: user
+    user: user._id
   })
   const savedBlog = await blog.save()
 
@@ -28,8 +34,21 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+
+  if (!request.token)
+    return response.status(401).json({ error: 'token missing' })
+
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if (!blog)
+    return response.status(400).json({ error: 'blog not found' })
+
+  const user = request.user
+  if (user.id !== blog.user.toString())
+    return response.status(401).json({ error: 'invalid token' })
+
   await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  return response.status(204).end()
 })
 
 blogRouter.put('/:id', async (request, response) => {
